Add files via upload
This first wave was subject to a set of revisions due to challenges with using internal libraries to load a predetermined config file.  After discussion, it was decided to use a mock config file to provide the test data we need for optimal code coverage and minimum reliance on external components.