Add lxml as a requirement
Whenever I try to follow the steps on xrpl-dev-portal to install Dactyl, the initial `dactyl_build` fails with the error `Couldn't find a tree builder with the features you requested: xml. Do you need to install a parser library?.` until I install `lxml` - I think that means that `lxml` should be in the requirements for the tool, although maybe I'm wrong.